[MPPI] complete minor optimaization with floating point calculations (#3827)
* floating point calculations

* Update optimizer_unit_tests.cpp

* Update critics_tests.cpp

* Update critics_tests.cpp

+++ nav2_mppi_controller/src/critics/path_align_critic.cpp
@@ -89,7 +89,7 @@ void PathAlignCritic::score(CriticData & data)
   }
 
   float dist_sq = 0.0f, dx = 0.0f, dy = 0.0f, dyaw = 0.0f, summed_dist = 0.0f;
-  double min_dist_sq = std::numeric_limits<float>::max();
+  float min_dist_sq = std::numeric_limits<float>::max();
   size_t min_s = 0;
 
   for (size_t t = 0; t < batch_size; ++t) {
